Making nuclei overall compatible with new changes + bug fixes

+++ v2/pkg/protocols/dns/dns.go
@@ -60,7 +60,7 @@ func (r *Request) Compile(options *protocols.ExecuterOptions) error {
 }
 
 // Requests returns the total number of requests the YAML rule will perform
-func (r *Request) Requests() int64 {
+func (r *Request) Requests() int {
 	return 1
 }
