+++ v2/pkg/protocols/dns/executer.go
@@ -30,8 +30,8 @@ func (e *Executer) Compile() error {
 }
 
 // Requests returns the total number of requests the rule will perform
-func (e *Executer) Requests() int64 {
-	var count int64
+func (e *Executer) Requests() int {
+	var count int
 	for _, request := range e.requests {
 		count += request.Requests()
 	}
